# CoreText: Faster OTC font loading