--- a/src/handle.rs
+++ b/src/handle.rs
@@ -14,11 +14,13 @@
 //!
 //! To open the font referenced by a handle, use a loader.
 
+use std::any::Any;
 use std::path::PathBuf;
 use std::sync::Arc;
 
 use crate::error::FontLoadingError;
 use crate::font::Font;
+use crate::loader::Loader;
 
 /// Encapsulates the information needed to locate and open a font.
 ///
@@ -45,6 +47,11 @@ pub enum Handle {
         /// If the memory consists of a single font, this value will be 0.
         font_index: u32,
     },
+    /// An already-loaded font.
+    Native {
+        /// Type-erased font storage. Use [`Self::as_native`] to retrieve the font object.
+        inner: Arc<dyn Any + Sync + Send>,
+    },
 }
 
 impl Handle {
@@ -66,6 +73,25 @@ impl Handle {
         Handle::Memory { bytes, font_index }
     }
 
+    /// Creates a new handle from a system handle.
+    pub fn from_native<T: Loader>(inner: &T) -> Self
+    where
+        T::NativeFont: Sync + Send,
+    {
+        Self::Native {
+            inner: Arc::new(inner.native_font()),
+        }
+    }
+    /// Retrieves a handle to the font object.
+    ///
+    /// May return None if inner object is not of type `T` or if this handle does not contain a native font object.
+    pub fn native_as<T: 'static>(&self) -> Option<&T> {
+        if let Self::Native { inner } = self {
+            inner.downcast_ref()
+        } else {
+            None
+        }
+    }
     /// A convenience method to load this handle with the default loader, producing a Font.
     #[inline]
     pub fn load(&self) -> Result<Font, FontLoadingError> {