--- a/tests/select_font.rs
+++ b/tests/select_font.rs
@@ -44,6 +44,7 @@ macro_rules! match_handle {
                     font_index, $index
                 );
             }
+            Handle::Native { .. } => {}
         }
     };
 }