Background: At [Zed](https://zed.dev) we've had users report that certain fonts could take a long time to load (https://github.com/zed-industries/community/issues/1745). Admittedly these font files are quite big (Iosevka font weights like 300Mb), but I've still went looking and found out that font-kit might sometimes copy excessive amount of memory, as seen in the flamegraph of loading of Iosevka:
![image](https://github.com/servo/font-kit/assets/24362066/7c291252-7814-4731-9340-b4e1121dd662)

This commit does several things:
- It uses native API to handle OTC via font descriptors. This gets rid of most of memmoves/memcopies
- It permits storing an already loaded font in Handle to avoid loading it in `load` again.
- It exposes `matching` module in public interface of a crate so that users don't have to go through Source::select_best_match`.

Note that I've pulled this commit verbatim from how we're using it in Zed; I made sure tests pass on Windows/Linux CI, but I didn't test it there (as we're only available on Mac for now). The macOS part works fine on production for few days already.

Another reason to have that commit (I guess) is that it vastly improves throughput of examples like `list-fonts` and `match-font` (https://github.com/servo/font-kit/issues/209). The particular command-line that this user was using runs 10x faster for me with this commit (0.15s vs 1.1s on M1 Max). `list-fonts` completes in ~2s on debug build (vs not completing within 10s or so on current master).